Comments for dynamic linking of patterns
Added a comment above the 'checkForConnectingPatterns' function explaining what it does, it will create two dictionaries that represent all of the following and preceding patterns that match other pattern names. These dictionaries are used in my template to dynamically create a link between the pattern itself and the pattern that is following and/or preceding the patter in question.